Fix the mypy gate: make a numpy-typed assignment unambiguous, and pin numpy
CI failed on the first run of the new static-analysis job:

  core/risk.py:628: error: Incompatible types in assignment (expression has
  type "floating[Any] | int", variable has type "float")  [assignment]

The same mypy version passed locally. The difference is numpy's bundled
stubs: np.mean() is typed differently across numpy releases, so the verdict
moved without the code moving. Two changes, because either alone leaves a
hole:

  - core/risk.py: var_95 and expected_shortfall are now wrapped in float()
    and default to 0.0 rather than 0. The values were always floats; the
    annotation and the expression just did not agree about it, and the mixed
    int/float defaults on the same lines made it worse.

  - Alpha IO/requirements-ci.txt: numpy pinned. A static gate whose result
    changes without the code changing is a gate people learn to ignore, and
    the checks should see a fixed stub baseline. requirements.txt still
    declares the range the application actually supports.

CHECKS ACTUALLY EXECUTED:
  MYPY: PASS - "no issues found in 19 source files".
  RUFF: PASS.
  TESTS: PASS - test_risk.py 26 passed, test_ci_gates.py 33 passed.

+++ Alpha IO/core/risk.py
@@ -624,8 +624,14 @@ class RiskManager:
             # VaR and Expected Shortfall
             sorted_returns = sorted(returns)
             var_index = int(len(sorted_returns) * 0.05)
-            metrics.var_95 = abs(sorted_returns[var_index]) if var_index < len(sorted_returns) else 0
-            metrics.expected_shortfall = abs(np.mean(sorted_returns[:var_index + 1])) if var_index > 0 else 0
+            metrics.var_95 = (
+                float(abs(sorted_returns[var_index]))
+                if var_index < len(sorted_returns) else 0.0
+            )
+            metrics.expected_shortfall = (
+                float(abs(np.mean(sorted_returns[:var_index + 1])))
+                if var_index > 0 else 0.0
+            )
 
         return metrics
